refactored and shifted header-file definitions
* shifted opencmw.hpp and related tests to core
* shifted Filters.hpp and related tests to core
* shifted QuerySerialiser.hpp and related tests to serialiser

* shifted fmt::format and operator<<(std::ostream) definitions to opencmw.hpp
  N.B. circular dependencies und sub-namespacing caused cognitive confusions
* TimingCtx: (temporarily) disabled serialising of 'bpcts' field
  N.B. need to refactor bpcts to Annotated/mp-units api
