+++ src/majordomo/test/CMakeLists.txt
@@ -63,8 +63,9 @@ opencmw_add_test_app(majordomo_testapp testapp.cpp)
 opencmw_add_test_app(majordomo_rest_server restserver_testapp.cpp)
 opencmw_add_test_app(majordomo_benchmark majordomo_benchmark.cpp)
 
+opencmw_add_test_catch2(RBAC_tests rbac_tests.cpp)
+opencmw_add_test_catch2(SubscriptionMatch_tests subscriptionmatcher_tests.cpp)
 opencmw_add_test_catch2(majordomo_tests majordomo_tests.cpp;subscriptionmatcher_tests.cpp)
 opencmw_add_test_catch2(majordomo_worker_tests majordomoworker_tests.cpp;subscriptionmatcher_tests.cpp)
 opencmw_add_test_catch2(majordomo_worker_rest_tests majordomoworker_rest_tests.cpp;subscriptionmatcher_tests.cpp)
-opencmw_add_test_catch2(queryserialiser_tests queryserialiser_tests.cpp)
