+++ src/majordomo/test/subscriptionmatcher_tests.cpp
@@ -1,18 +1,18 @@
 #include <catch2/catch.hpp>
 
-#include <majordomo/Filters.hpp>
+#include <Filters.hpp>
 #include <majordomo/SubscriptionMatcher.hpp>
-
 #include <TimingCtx.hpp>
+#include <QuerySerialiser.hpp>
+
 #include <URI.hpp>
 
 #include <charconv>
 
-using opencmw::majordomo::DomainFilter;
 using opencmw::majordomo::SubscriptionMatcher;
-using URI = SubscriptionMatcher::URI;
 
 TEST_CASE("Test path-only topics", "[subscription_matcher][path_only]") {
+    using URI = SubscriptionMatcher::URI;
     SubscriptionMatcher matcher;
 
     REQUIRE(matcher(URI("/property"), URI("/property")));
@@ -51,16 +51,15 @@ struct Int {
     bool operator==(const Int &) const = default;
 };
 
-using TestFilter1 = DomainFilter<Int>;
-using TestFilter2 = DomainFilter<std::string_view>;
-
-struct LessThan {
-    bool operator()(std::string_view lhs, std::string_view rhs) const { return lhs < rhs; }
-};
-
-using TestFilter3 = DomainFilter<std::string_view, LessThan>;
-
 TEST_CASE("Test path and query", "[subscription_matcher][path_and_query]") {
+    struct LessThan {
+        bool operator()(std::string_view lhs, std::string_view rhs) const { return lhs < rhs; }
+    };
+    using opencmw::DomainFilter;
+    using TestFilter1 = DomainFilter<Int>;
+    using TestFilter2 = DomainFilter<std::string_view>;
+    using TestFilter3 = DomainFilter<std::string_view, LessThan>;
+    using URI         = SubscriptionMatcher::URI;
     SubscriptionMatcher matcher;
     matcher.addFilter<TestFilter1>("testKey1");
     matcher.addFilter<TestFilter2>("testKey2");
@@ -84,9 +83,10 @@ TEST_CASE("Test path and query", "[subscription_matcher][path_and_query]") {
 }
 
 TEST_CASE("Test timing and context type matching", "[subscription_matcher]") {
+    using URI = SubscriptionMatcher::URI;
     SubscriptionMatcher matcher;
-    matcher.addFilter<opencmw::majordomo::TimingCtxFilter>("ctx");
-    matcher.addFilter<opencmw::majordomo::ContentTypeFilter>("contentType");
+    matcher.addFilter<opencmw::TimingCtxFilter>("ctx");
+    matcher.addFilter<opencmw::ContentTypeFilter>("contentType");
 
     REQUIRE_FALSE(matcher(URI("/property?ctx=FAIR.SELECTOR.ALL"), URI("/property?ctx=FAIR.SELECTOR.C=2")));
     REQUIRE(matcher(URI("/property?ctx=FAIR.SELECTOR.C=2"), URI("/property?ctx=FAIR.SELECTOR.ALL")));
@@ -94,6 +94,26 @@ TEST_CASE("Test timing and context type matching", "[subscription_matcher]") {
     REQUIRE(matcher(URI("/property?ctx=FAIR.SELECTOR.C=2:P=1"), URI("/property?ctx=FAIR.SELECTOR.C=2")));
     REQUIRE(matcher(URI("/property?ctx=FAIR.SELECTOR.C=2:P=1"), URI("/property?ctx=FAIR.SELECTOR.C=2:P=1")));
     REQUIRE_FALSE(matcher(URI("/property?ctx=FAIR.SELECTOR.C=2"), URI("/property?ctx=FAIR.SELECTOR.C=2:P=1"))); // notify not specific enough (missing 'P=1')
+    REQUIRE_FALSE(matcher(URI("/property?ctx=FAIR.SELECTOR.ALL&contentType=text/html"), URI("/property?ctx=FAIR.SELECTOR.C=2&contentType=text/html")));
+    REQUIRE_FALSE(matcher(URI("/property?ctx=FAIR.SELECTOR.ALL"), URI("/property?ctx=FAIR.SELECTOR.C=2&contentType=text/html")));
+}
+
+struct MatcherTest {
+    opencmw::TimingCtx      ctx;
+    opencmw::MIME::MimeType contentType = opencmw::MIME::UNKNOWN;
+};
+ENABLE_REFLECTION_FOR(MatcherTest, ctx, contentType)
+
+TEST_CASE("Register filters", "[QuerySerialiser][register_filters]") {
+    using URI = opencmw::URI<opencmw::RELAXED>;
+
+    opencmw::majordomo::SubscriptionMatcher matcher;
+    opencmw::query::registerTypes(MatcherTest(), matcher);
+
+    // subset of the subscriptionfilter_tests
+    REQUIRE_FALSE(matcher(URI("/property?ctx=FAIR.SELECTOR.ALL"), URI("/property?ctx=FAIR.SELECTOR.C=2")));
+    REQUIRE(matcher(URI("/property?ctx=FAIR.SELECTOR.C=2"), URI("/property?ctx=FAIR.SELECTOR.ALL")));
+    REQUIRE(matcher(URI("/property?ctx=FAIR.SELECTOR.C=2"), URI("/property?ctx=FAIR.SELECTOR.C=2")));
     REQUIRE_FALSE(matcher(URI("/property?ctx=FAIR.SELECTOR.ALL&contentType=text/html"), URI("/property?ctx=FAIR.SELECTOR.C=2&contentType=text/html")));
     REQUIRE_FALSE(matcher(URI("/property?ctx=FAIR.SELECTOR.ALL"), URI("/property?ctx=FAIR.SELECTOR.C=2&contentType=text/html")));
 }
